Moved help functions into utils_test.py
