+++ ucp/_libs/tests/test_peer_send_recv.py
@@ -1,47 +1,12 @@
 import multiprocessing as mp
 import os
-import time
 
 import pytest
 
 from ucp._libs import ucx_api
-from ucp._libs.arr import Array
+from ucp._libs.utils_test import blocking_recv, blocking_send
 
 mp = mp.get_context("spawn")
-
-
-def blocking_handler(request, exception, finished):
-    finished[0] = True
-
-
-def blocking_send(worker, ep, msg, tag):
-    msg = Array(msg)
-    finished = [False]
-    req = ucx_api.tag_send_nb(
-        ep, msg, msg.nbytes, tag=tag, cb_func=blocking_handler, cb_args=(finished,),
-    )
-    if req is not None:
-        while not finished[0]:
-            worker.progress()
-            time.sleep(0.1)
-
-
-def blocking_recv(worker, ep, msg, tag):
-    msg = Array(msg)
-    finished = [False]
-    req = ucx_api.tag_recv_nb(
-        worker,
-        msg,
-        msg.nbytes,
-        tag=tag,
-        cb_func=blocking_handler,
-        cb_args=(finished,),
-        ep=ep,
-    )
-    if req is not None:
-        while not finished[0]:
-            worker.progress()
-            time.sleep(0.1)
 
 
 def _test_peer_communication(queue, rank, msg_size):
